requirements.txt
fastapi>=0.100.0
sqlmodel>=0.0.8
uvicorn>=0.22.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0  # Agregado para manejar configuraciones
python-multipart>=0.0.6
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.1
pytest-cov>=4.1.0
colorama>=0.4.6
python-json-logger>=2.0.7